add additional permission checks in homescript

+++ core/homescript/executor.go
@@ -3,7 +3,10 @@ package homescript
 import (
 	"errors"
 	"fmt"
+	"strings"
 	"time"
+
+	"golang.org/x/exp/utf8string"
 
 	"github.com/MikMuellerDev/homescript/homescript/interpreter"
 	"github.com/MikMuellerDev/smarthome/core/database"
@@ -18,7 +21,9 @@ type Executor struct {
 	Output     string
 }
 
-// Prints to the console
+// Emulates printing to the console
+// Instead, appends the provided message to the output of the executor
+// Exists in order to return the script's output to the user
 func (self *Executor) Print(args ...string) {
 	var output string
 	for _, arg := range args {
@@ -28,6 +33,7 @@ func (self *Executor) Print(args ...string) {
 }
 
 // Returns a boolean if the requested switch is on or off
+// Returns an error if the provided switch does not exist
 func (self *Executor) SwitchOn(switchId string) (bool, error) {
 	powerState, err := hardware.GetPowerState(switchId)
 	if err != nil {
@@ -36,7 +42,9 @@ func (self *Executor) SwitchOn(switchId string) (bool, error) {
 	return powerState, err
 }
 
-// Changes the power state on said switch
+// Changes the power state of an arbitrary switch
+// Checks if the switch exists, if the user is allowed to interact with switches and if the user has the matching switch-permission
+// If a check fails, an error is returned
 func (self *Executor) Switch(switchId string, powerOn bool) error {
 	err := hardware.SetSwitchPowerAll(switchId, powerOn, self.Username)
 	if err != nil {
@@ -52,11 +60,12 @@ func (self *Executor) Switch(switchId string, powerOn bool) error {
 }
 
 // Sends a mode request to a given radiGo server
+// TODO: implement this feature
 func (self *Executor) Play(server string, mode string) error {
 	return errors.New("The feature 'radiGo' is not yet implemented")
 }
 
-// Sends a notification to the current user
+// Sends a notification to the user who issues this command
 func (self *Executor) Notify(
 	title string,
 	description string,
@@ -74,44 +83,98 @@ func (self *Executor) Notify(
 	return nil
 }
 
+// Adds a new user to the system
+// If the user already exists, an error is returned
 func (self *Executor) AddUser(username string, password string, forename string, surname string) error {
+	hasPermission, err := database.UserHasPermission(self.Username, database.PermissionManageUsers)
+	if err != nil {
+		return fmt.Errorf("Failed to add user: could not validate your permissions: %s", err.Error())
+	}
+	if !hasPermission {
+		return errors.New("Failed to add user: You lack permission to manage users: if this is unintentional, contact your administrator.")
+	}
+	if len(username) == 0 || strings.Contains(username, " ") || !utf8string.NewString(username).IsASCII() {
+		return errors.New("Failed to add user: username should only include ASCII characters and must not contain whitespaces or be blank")
+	}
+	if len(password) == 0 {
+		return errors.New("Blank passwords are not allowed")
+	}
+	if len(username) > 20 || len(forename) > 20 || len(surname) > 20 {
+		return errors.New("length of username, forename, and surname must not exceed 20 characters")
+	}
 	if err := database.AddUser(database.FullUser{
 		Username: username,
 		Password: password,
 		Forename: forename,
 		Surname:  surname,
 	}); err != nil {
-		log.Error(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': failed to create user: %s", self.ScriptName, username, err.Error()))
 		return err
 	}
 	return nil
 }
 
+// Deletes
 func (self *Executor) DelUser(username string) error {
+	hasPermission, err := database.UserHasPermission(self.Username, database.PermissionManageUsers)
+	if err != nil {
+		return fmt.Errorf("Failed to remove user: could not validate your permissions: %s", err.Error())
+	}
+	if !hasPermission {
+		return errors.New("Failed to remove user: You lack permission to manage users: if this is unintentional, contact your administrator.")
+	}
+	isStandaloneUserAdmin, err := user.IsStandaloneUserAdmin(username)
+	if err != nil {
+		return err
+	}
+	if isStandaloneUserAdmin {
+		return errors.New("Did not delete user: target is the only user-administrator: deleting this user would break the system.")
+	}
 	if err := user.DeleteUser(username); err != nil {
-		log.Error(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': failed to delete user: %s", self.ScriptName, username, err.Error()))
 		return err
 	}
 	return nil
 }
 
 func (self *Executor) AddPerm(username string, permission string) error {
+	hasPermission, err := database.UserHasPermission(self.Username, database.PermissionManageUsers)
+	if err != nil {
+		return fmt.Errorf("Failed to remove user: could not validate your permissions: %s", err.Error())
+	}
+	if !hasPermission {
+		return errors.New("Failed to remove user: You lack permission to manage users: if this is unintentional, contact your administrator.")
+	}
 	if !database.DoesPermissionExist(permission) {
-		return fmt.Errorf("The permission '%s' does not exist.", permission)
+		return fmt.Errorf("Failed to add permission: the permission '%s' does not exist. You can view a complete list of valid permissions under user-management > manage user permissions (any user) > permissions", permission)
 	}
 	edited, err := user.AddPermission(username, database.PermissionType(permission))
 	if err != nil {
 		return fmt.Errorf("Failed to add permission: database failure: %s", err.Error())
 	}
 	if !edited {
-		return errors.New("User already has this permission")
+		return errors.New("Did not add permission: user already has this permission")
 	}
 	return nil
 }
 
 func (self *Executor) DelPerm(username string, permission string) error {
+	hasPermission, err := database.UserHasPermission(self.Username, database.PermissionManageUsers)
+	if err != nil {
+		return fmt.Errorf("Failed to remove user: could not validate your permissions: %s", err.Error())
+	}
+	if !hasPermission {
+		return errors.New("Failed to remove user: You lack permission to manage users: if this is unintentional, contact your administrator.")
+	}
 	if !database.DoesPermissionExist(permission) {
 		return fmt.Errorf("The permission '%s' does not exist.", permission)
+	}
+	if permission == string(database.PermissionManageUsers) || permission == string(database.PermissionWildCard) {
+		isAlone, err := user.IsStandaloneUserAdmin(username)
+		if err != nil {
+			return err
+		}
+		if isAlone {
+			return errors.New("Did not remove permission: target user is the only user-administrator: removing this permission would break the system.")
+		}
 	}
 	edited, err := user.RemovePermission(username, database.PermissionType(permission))
 	if err != nil {
@@ -131,11 +194,9 @@ func (self *Executor) Log(
 ) error {
 	hasPermission, err := database.UserHasPermission(self.Username, database.PermissionLogs)
 	if err != nil {
-		log.Error(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': failed to log event: failed to check permission: %s", self.ScriptName, self.Username, err.Error()))
 		return err
 	}
 	if !hasPermission {
-		log.Error(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': failed to log event: failed to check permission: %s", self.ScriptName, self.Username, err.Error()))
 		return fmt.Errorf("Failed to add log event: user '%s' is not allowed to use the internal logging system.", self.Username)
 	}
 	switch level {
@@ -152,7 +213,6 @@ func (self *Executor) Log(
 	case 5:
 		event.Fatal(title, description)
 	default:
-		log.Debug(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': failed to log event: invalid level", self.ScriptName, self.Username))
 		return fmt.Errorf("Failed to add log event: invalid logging level <%d>: valid logging levels are 1, 2, 3, 4, or 5", level)
 	}
 	return nil
@@ -162,7 +222,7 @@ func (self *Executor) Log(
 func (self Executor) Exec(homescriptId string) (string, error) {
 	output, exitCode, err := RunById(self.Username, homescriptId)
 	if err != nil {
-		log.Debug(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': called homescript failed with exit code %d : %s", self.ScriptName, self.Username, exitCode, err.Error()))
+		self.Print(fmt.Sprintf("Exec failed: called homescript failed with exit code %d", exitCode))
 		return output, err
 	}
 	return output, nil
@@ -175,13 +235,11 @@ func (self *Executor) GetUser() string {
 
 // TODO: Will later be implemented, should return the weather as a human-readable string
 func (self *Executor) GetWeather() (string, error) {
-	log.Warn(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': weather is not implemented yet", self.ScriptName, self.Username))
 	return "rainy", nil
 }
 
 // TODO: Will later be implemented, should return the temperature in Celsius
 func (self *Executor) GetTemperature() (int, error) {
-	log.Warn(fmt.Sprintf("[Homescript] ERROR: script: '%s' user: '%s': temperature is not implemented yet", self.ScriptName, self.Username))
 	return 42, nil
 }
